feat(protocol-designer): refactor and restyle LabwareSelectionModal
Closes #1781
